FastAPI/requirements.txt
astroid==3.2.4
black==24.8.0
click==8.1.7
dill==0.3.8
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.2
pylint==3.2.7
tomlkit==0.13.2

annotated-types==0.7.0
anyio==4.4.0
fastapi==0.112.1
h11==0.14.0
idna==3.7
mysql==0.0.3
mysqlclient==2.2.4
peewee==3.17.6
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6